Add checking the status of buffer queue
This patch adds checking the status of buffer queue
to finish the thread after dequeing all buffers
because it can send a empty buffer if the thread is finished
before dequeing all buffers.
and changes not to stop running stream threads
if the sensor setting of the streams is the same
with a new stream because of a performance issue.

Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>
Change-Id: Ic64b56326e4bf1016c99d50e804f85fce87ab6ab
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/android_hal_camera/+/456438
Reviewed-by: Sungwoo Park <swpark@nexell.co.kr>
Tested-by: Sungwoo Park <swpark@nexell.co.kr>